requirements.txt
Flask==2.3.3
mysql-connector-python==8.2.0
Werkzeug==2.3.7
PyJWT==2.12.1
pytest==7.4.3
pytest-cov==4.1.0